update autoclean trace
