+++ general/quit.py
@@ -1,5 +1,6 @@
 from libs.config import alias, gset
 from libs.app import sys_exit
+from libs.myapp import clean_trace
 
 
 # ? alias装饰器第一个参数为none_named_arg(true/FALSE),True即把没有参数名时传入的参数值顺序传入
@@ -12,4 +13,5 @@ def run():
     Quit this program.
     """
     gset("loop", False, True)
+    clean_trace()
     sys_exit()
