+++ webshell_plugins/back.py
@@ -1,14 +1,5 @@
-from libs.config import alias, gget, gset, color, set_namespace
-from libs.myapp import send, get_system_code, is_windows, base64_encode
-
-
-def get_clean_ld_preload_php(filename: str):
-    system_clean_command = f"del /F /Q {filename}" if is_windows() else f"rm -f {filename}" + " && echo success"
-    return """$f=base64_decode("%s");
-if (!unlink($f)){
-    %s
-}else{echo "success";}
-""" % (base64_encode(filename), get_system_code(system_clean_command))
+from libs.config import alias, set_namespace
+from libs.myapp import clean_trace
 
 
 # ? alias装饰器第一个参数为none_named_arg(true/FALSE),True即把没有参数名时传入的参数值顺序传入
@@ -19,16 +10,5 @@ def run():
 
     Back to main menu.
     """
-    ld_preload_filename = gget("webshell.ld_preload_path", "webshell", None)
-    if (ld_preload_filename):
-        print(color.yellow("\nClean LD_PRELOAD traces...\n"))
-        res = send(get_clean_ld_preload_php(ld_preload_filename))
-        if (res):
-            text = res.r_text.strip()
-            if ("success" in text):
-                print(color.green("Clean success\n"))
-            else:
-                print(color.red("Clean failed\n"))
-    gset("webshell.ld_preload_path", None, True, "webshell")
-    gset("webshell.ld_preload_func", None, True, "webshell")
+    clean_trace()
     set_namespace("main")
